Update Makefile to use dep rather than glide
Now that we've moved from glide to dep, we remove go getting glide and
running glide install. Rather, we use dep ensure (in verbose) to resolve
our dependencies as calcified in .lock and as configured in .toml.
Fortunately, due to the similarities in how glide and dep operate, the
changes are minor.